refactor(dashboard): extract page entry components, thin routes
